create bloc provider and bloc listener and bloc observer and add note to data base

+++ lib/models/note_model.dart
@@ -10,7 +10,7 @@ class NoteModel extends HiveObject
   @HiveField(2)
   final String date;
   @HiveField(3)
-  final int color;
+  final double color;
 
   NoteModel({required this.title,required this.subTitle,required this.date,required this.color});
 }
